sma_pipeline/requirements.txt
# Core
numpy
pandas
scipy

# Config handling
pyyaml

# Plotting
matplotlib
seaborn
